Added code that will handle the view models and world models for weapon skin changes. Added militia model. Added brownbess_nobayo Added rev back in. Slowed skirmishers down a tad. Added a CTF Flag hud (Not quite finished) Comm menus will no longer load if you are not on a team or dead.
Spectator menu removed, it was useless.
Removed a useless Team Menu (Not related to BG2)
Huds are no longer drawn if the entites they are connected to no longer exist.

If you holster while reloading, you will loose a shot. (This may be changed later)

Native now spawns with the brownbess_nobayo as well as the tomahawk.
Militia spawns with the rev.