Make ironsights toggle based on the state of IN_ZOOM (+zoom and -zoom) and make the "ironsights" command toggle IN_ZOOM. Also fix ironsights handling on the client.
Basically, EnableIronsights()/DisableIronsights() are called based on whether IN_ZOOM is set in m_nButtons.
This fixes a lot of quick scoping related client glitches, such as lack of effects of spurious effects.
The interpolation factor in baseviewmodel_shared.cpp could sometimes get out of range due to this, so that also required fixing.
A nice side effect of all this is that the player is able to bind +zoom to a button and obtain ironsights while *holding* the button instead of toggling. The old way still works though.

/Tjoppen