Fix bayonet retrace duration bug introduced in r565.
Using ATTACK_PRIMARY's retrace duration works fine for slashing weapons since both attacks are identical.
For bayonets this led to a duration of zero, hence no retracing.
This was basically a copy-paste bug - the next commit will refactor CBaseBG2Weapon::*Attack() into a common function to avoid this in the future.

/Tjoppen